+++ changelog.md
@@ -19,6 +19,7 @@
     - Change the trim side on the fly by pressing the letter <kbd>O</kbd>
     - Toggle ripple edit and select mode on the fly with <kbd>Alt</kbd> and <kbd>Shift</kbd> respectively
 - `Import Local Footage` does not impose a folder structure on you anymore
+- `Set Playback Speed`: Add shortcut to set the speed to 3x
 
 ### Changes ###